+++ src/utils/chatFormat.js
@@ -1,3 +1,30 @@
+export const formatChatTime = (timestamp) => {
+  if (!timestamp) return '';
+  const date = timestamp.toDate();
+  const now = new Date();
+  const diffMs = now - date;
+
+  if (diffMs < 60 * 1000) return '방금';
+
+  const todayMidnight = new Date(now.getFullYear(), now.getMonth(), now.getDate());
+  if (date >= todayMidnight) {
+    return date.toLocaleTimeString('ko-KR', { hour: '2-digit', minute: '2-digit', hour12: false });
+  }
+
+  const yesterdayMidnight = new Date(todayMidnight);
+  yesterdayMidnight.setDate(yesterdayMidnight.getDate() - 1);
+  if (date >= yesterdayMidnight) return '어제';
+
+  if (date.getFullYear() === now.getFullYear()) {
+    return `${date.getMonth() + 1}월 ${date.getDate()}일`;
+  }
+
+  const y = date.getFullYear();
+  const m = String(date.getMonth() + 1).padStart(2, '0');
+  const d = String(date.getDate()).padStart(2, '0');
+  return `${y}.${m}.${d}`;
+};
+
 export const formatMsgTime = (timestamp) => {
   if (!timestamp) return '';
   const date = timestamp.toDate();
